Update README.md
Hi,

I have two suggestions for this file. These are:

"The deployment can be done using file installed..." should be "The deployment can be done using a file installed..." for grammatical correctness and,

"The Gevulot node try to open..." should be "The Gevulot node tries to open..." for grammatical correctness.

Thanks.